feat(playback): add MP3 index seeking option
This adds a new experimental option to enable index-based seeking for MP3 files. When enabled, the entire MP3 file is loaded into memory to build a seek index, which can fix seeking issues for some VBR files that lack a proper Xing header.

The option is disabled by default and can be found in the playback settings. Enabling it requires an app restart to take effect.