Feat: move cover transitions and more
### Summary

Moved the album cover transitions to the Now Playing styles dialog.
This also restricts certain transitions to certain styles that look
best with each other. May need more refining there.

Also, removed the existing transitions arrays and dialog in settings
page, since it was moved. Set the carousel effect to off by default
for now, as it causes some issues with the transitions.